requirements.txt
# Core dependencies
python-dotenv>=0.21.0
discord.py>=2.3.2
openai>=1.0.0
aiohttp>=3.8.1
requests>=2.28.0

# Additional dependencies
tiktoken>=0.4.0   
PyNaCl>=1.5.0      
